Playing with an if
So I know there must be a better way to do this, but this if statement looks to be pulsing the const bool values. at each interrupt with back to back zeros there is a blip though. Not sure this matters though since the 38khz receiver would still see those big blips as a zero.